requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==2.7.0
Flask-MQTT==1.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
fonttools==4.47.2
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==1.4.6
numpy==1.26.3
packaging==23.2
paho-mqtt==1.5.1
pillow==10.2.0
ply==3.11
protobuf==4.21.12
pyasn1==0.5.1
pycryptodomex==3.19.1
pyparsing==3.1.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
